pyproject.toml
[project]
name = "roll-tracker"
version = "0.0.1"
description = "Offline BJJ video pipeline (skeleton)"
authors = [{ name = "Bryan Thomas" }]
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2,<3",
  "pandas>=2,<3",
  "typer>=0.9,<1",
  "pyarrow>=14,<16",
  "pyyaml>=6,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "opencv-python>=4.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bjj_pipeline"]

[tool.hatch.build.targets.sdist]
include = ["src/bjj_pipeline"]
